Feat: Create sortable module and spec
