+++ spec/league_spec.rb
@@ -44,10 +44,6 @@ describe League do
     expect(@stat_tracker.league.team_id_to_team_name("4")).to eq("Chicago Fire")
   end
 
-  it '#game_team_group_by_season' do
-    expect(@stat_tracker.league.game_team_group_by_season("20122013")[2].game_id).to eq("2012030222")
-  end
-
   it '#coaches_by_win_percentage' do
     expect(@stat_tracker.league.coaches_by_win_percentage("20122013")).to eq({"John Tortorella" => 0.00,
                   "Claude Julien" => 100.00})
@@ -70,10 +66,6 @@ describe League do
       "link" => "/api/v1/teams/26"
     }
     expect(@stat_tracker.team_info("26")).to eq(expected)
-  end
-
-  it '#game_team_group_by_team' do
-    expect(@stat_tracker.league.game_team_grouped_by_team("3").length).to eq(2)
   end
 
   it '#data_sorted_by_season' do
@@ -130,7 +122,7 @@ describe League do
 
     expect(@stat_tracker.league.win_percentage_by_opponent("28")).to eq({"17" => 0.0, "18" => 100.0,})
     expect(@stat_tracker.league.win_percentage_by_opponent("1")).to eq({"25" => 100.0, "4" => 0.0})
-  end 
+  end
 
   it '#home_wins_counter' do
     expect(@stat_tracker.league.home_wins_counter).to eq(1)
@@ -144,24 +136,12 @@ describe League do
     expect(@stat_tracker.league.tie_counter).to eq(2)
   end
 
-  it '#goals_by_season' do
-    expect(@stat_tracker.league.goals_by_season).to eq({"20132014" => [1], "20142015" => [6,1], "20152016" => [6,5]})
-  end
-
   it '#avg_goals_by_season' do
     expect(@stat_tracker.league.avg_goals_by_season).to eq({"20132014" => 1, "20142015" => 3.5, "20152016" => 5.5})
   end
 
-  it '#away_team_by_goals' do
-    expect(@stat_tracker.league.away_team_by_goals).to eq({"3" => [2, 2], "6" => [2]})
-  end
-
   it '#avg_away_team_by_goals' do
     expect(@stat_tracker.league.avg_away_team_by_goals).to eq({"3" => 2, "6" => 2})
-  end
-
-  it '#home_team_by_goals' do
-    expect(@stat_tracker.league.home_team_by_goals).to eq({"6" => [3, 3]})
   end
 
   it '#avg_home_team_by_goals' do
